requirements.txt
mss==9.0.1
numpy==1.23.4
Pillow==9.4.0
Pillow==10.3.0
pyautogui==0.9.54
